+++ frameworks/preact/src/types/field.ts
@@ -79,6 +79,12 @@ export interface FieldStore<
    */
   readonly isValid: ReadonlySignal<boolean>;
   /**
+   * Sets the field input value programmatically.
+   */
+  readonly onInput: (
+    value: PartialValues<PathValue<v.InferInput<TSchema>, TFieldPath>>
+  ) => void;
+  /**
    * The props to spread onto the field element for integration.
    */
   readonly props: FieldElementProps;