+++ frameworks/qwik/src/hooks/useField/useField.ts
@@ -82,6 +82,10 @@ export function useField(form: FormStore, config: UseFieldConfig): FieldStore {
     isValid: createComputed$(
       () => !getFieldBool(internalFieldStore.value, 'errors')
     ),
+    onInput: $((value) => {
+      setFieldInput(internalFormStore, pathSignal.value, value);
+      validateIfRequired(internalFormStore, internalFieldStore.value, 'input');
+    }),
     props: {
       get name() {
         return internalFieldStore.value.name;