+++ frameworks/qwik/src/types/field.ts
@@ -78,6 +78,12 @@ export interface FieldStore<
    */
   readonly isValid: ReadonlySignal<boolean>;
   /**
+   * Sets the field input value programmatically.
+   */
+  readonly onInput: QRL<
+    (value: PartialValues<PathValue<v.InferInput<TSchema>, TFieldPath>>) => void
+  >;
+  /**
    * The props to spread onto the field element for integration.
    */
   readonly props: FieldElementProps;