+++ frameworks/react/src/types/field.ts
@@ -72,6 +72,12 @@ export interface FieldStore<
    */
   readonly isValid: boolean;
   /**
+   * Sets the field input value programmatically.
+   */
+  readonly onChange: (
+    value: PartialValues<PathValue<v.InferInput<TSchema>, TFieldPath>>
+  ) => void;
+  /**
    * The props to spread onto the field element for integration.
    */
   readonly props: FieldElementProps;